\name{C++Class-class}
\Rdversion{1.1}
\docType{class}
\alias{C++Class-class}
\alias{new,C++Class-method}

\title{Class "C++Class"}
\description{C++ class}
\section{Objects from the Class}{
	Objects can be created by calling the \link{CPP} 
	function
}
\section{Slots}{
  \describe{
    \item{\code{cppclass}:}{name of the associated C++ class}
  }
}
\section{Methods}{
  \describe{
    \item{new}{\code{signature(Class = "C++Class")}: creates a new object of the C++ class }
	 }
}
\author{ Romain Francois <francoisromain@free.fr> }
\seealso{\link{CPP}}
\keyword{classes}
