\name{C++Object-class}
\Rdversion{1.1}
\docType{class}
\alias{C++Object-class}
\alias{$,C++Object-method}

\title{Virtual Class "C++Object" }
\description{Virtual class, super class of all C++ objects}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Slots}{
  \describe{
    \item{\code{pointer}:}{external pointer to the actual C++ object}
  }
}
\section{Methods}{
  \describe{
    \item{$}{\code{signature(x = "C++Object")}: invokes a method of the C++ object }
	 }
}
\author{ Romain Francois <francoisromain@free.fr> }
\keyword{classes}
