\name{CPP-package}
\alias{CPP-package}
\alias{CPP}
\docType{package}
\title{
A Package for the Composition of Probabilistic Preferences (CPP)
}
\description{
This package supports the decision making in multicriteria problems. The CPP method uses a probabilistic approach to emulate the uncertainty that is inevitably present in preference evaluations. This uncertainty is modelled by probability distributions around the exact measurements of alternatives in all criteria considered in the problem. This package includes several functions described in the references, with a focus on ranking, sorting and analysing the evolution of alternatives.
}
\author{
Luiz Octavio Gaviao (luiz.gaviao67@gmail.com);

Annibal Parracho Sant'Anna (annibal.parracho@gmail.com);

Gilson Brito Alves Lima (glima@id.uff.br); and

Pauli Adriano de Almada Garcia (pauliadriano@gmail.com).
}
\references{
Sant'Anna, Annibal P. (2015). Probabilistic Composition of Preferences: Theory and Applications, Springer.

Gaviao, Luiz O. & Lima, Gilson B.A. (2017) Support decision to player selection: an application of the CPP in soccer, Novas Edicoes Academicas [in Portuguese].
}
