\name{CPP}
\alias{CPP}
\title{
C++ class resolver
}
\description{
	Creates a \linkS4class{C++Class} object that can be used 
	to build a \linkS4class{C++Object} of the given class
}
\usage{
CPP(cppclass)
}
\arguments{
  \item{cppclass}{The name of the C++ class}
}
\value{
	a \linkS4class{C++Class} object
}
\author{
Romain Francois <francoisromain@free.fr>
}
\keyword{interface}
