% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r-package.R
\docType{package}
\name{cpp4r-package}
\alias{cpp4r}
\alias{cpp4r-package}
\title{cpp4r: A C++ Interface for R's C Interface}
\description{
Provides a header only, C++ interface to R's C interface. Compared to other approaches 'cpp4r' strives to be safe against long jumps from the C API as well as C++ exceptions, conform to normal R function semantics and supports interaction with 'ALTREP' vectors.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cpp4r.org}
  \item \url{https://github.com/pachadotdev/cpp4r}
  \item Report bugs at \url{https://github.com/pachadotdev/cpp4r/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.vargas.sepulveda@gmail.com} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

Authors:
\itemize{
  \item  Posit Software, PBC (Original cpp11 package)
}

}
\keyword{internal}
