\name{PACERcpp-package}
\alias{PACERcpp-package}
\alias{PACERcpp}
\docType{package}
\title{
\packageTitle{PACERcpp}
}
\description{
\packageDescription{PACERcpp}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PACERcpp}
\packageIndices{PACERcpp}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{PACERcpp}

Maintainer: \packageMaintainer{PACERcpp}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
