\name{set<int>-class}
\Rdversion{1.1}
\docType{class}
\alias{set<int>-class}

\title{Class "set<int>" ~~~ }
\description{wrapper around the \code{std::set<int>} c++ class}
\section{Objects from the Class}{
Objects can be created by calls of the form 
\code{new( CPP("set<int>" ) )}.
}
\section{Slots}{
  \describe{
  	  \item{\code{pointer}:}{external pointer to the C++ \code{std::set<int>} object }
  }
}
\section{Extends}{
Class \code{"\linkS4class{set<>}"}, directly.
Class \code{"\linkS4class{C++Object}"}, by class "set<>", distance 2.
}
\section{Methods}{
No methods defined with class "set<int>" in the signature.
}
\references{ The \code{std::set} C++ template from the standard template library.  }
\author{ Romain Francois <francoisromain@free.fr> }
\keyword{classes}
