% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unvendor.R
\name{unvendor}
\alias{unvendor}
\title{Unvendor the cpp4r headers}
\usage{
unvendor(path = ".")
}
\arguments{
\item{path}{The path to the package root directory}
}
\value{
The file path to the unvendored code (invisibly).
}
\description{
This function removes the vendored cpp4r headers from your package and
restores the \code{LinkingTo: cpp4r} field in the DESCRIPTION file if it was removed.
}
\examples{
# create a new directory
dir <- tempfile()
dir.create(dir)

# vendor the cpp4r headers into the directory
vendor(dir)

# unvendor the cpp4r headers from the directory
unvendor(dir)

list.files(file.path(dir, "inst", "include", "cpp4r"))

# cleanup
unlink(dir, recursive = TRUE)
}
