\name{CPP-package}
\alias{CPP-package}
\docType{package}
\title{
C++ objects in R
}
\description{
The package defines a simple mechanism to expose 
C++ classes as R (S4 classes)
}
\details{
\tabular{ll}{
Package: \tab CPP\cr
Type: \tab Package\cr
Version: \tab 0.0-1\cr
Date: \tab 2010-01-22\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Romain Francois

Maintainer: Romain Francois <francoisromain@free.fr>
}
\keyword{ package }

