% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r-package.R
\docType{package}
\name{cpp4r-package}
\alias{cpp4r}
\alias{cpp4r-package}
\title{cpp4r: Header-Only 'C++' and 'R' Interface}
\description{
Provides a header only, 'C++' interface to 'R' with enhancements over 'cpp11'. Enforces copy-on-write semantics consistent with 'R' behavior. Offers native support for ALTREP objects, 'UTF-8' string handling, modern 'C++11' features and idioms, and reduced memory requirements. Allows for vendoring, making it useful for restricted environments. Compared to 'cpp11', it adds support for converting 'C++' maps to 'R' lists, 'Roxygen' documentation directly in 'C++' code, proper handling of matrix attributes, support for nullable external pointers, bidirectional copy of complex number types, flexibility in type conversions, use of nullable pointers, and various performance optimizations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cpp4r.org}
  \item \url{https://github.com/pachadotdev/cpp4r}
  \item Report bugs at \url{https://github.com/pachadotdev/cpp4r/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.vargas.sepulveda@gmail.com} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

Authors:
\itemize{
  \item  Posit Software, PBC (Original cpp11 package)
}

}
\keyword{internal}
